SCC API check: move
