+++ pkg/security/scc/bypriority_test.go
@@ -1,4 +1,4 @@
-package admission
+package scc
 
 import (
 	"sort"
